I've managed to get the Dockerfile to a point where you can run it as a non-root user.
Quickstart instructions:
- docker build -t follow -t follow:latest .
- xhost +local:nobody
- docker run -it -d --rm --network host follow:latest

note: if you are using --userns-remap this will not work for you,
alternatively you can run a separate docker daemon without if you
need to retain that functionality and want to to use this.
For more info:
https://docs.docker.com/engine/reference/commandline/dockerd/#run-multiple-daemons

❯ docker build -t follow -t follow:latest .
❯ xhost local:root
❯ docker run -it --rm --userns host --network host follow:latest

Dockerfile/.dockerignore for building and running a dev env (works
with X11.) There certainly probably are better ways to to export
a uid/gid to the container than to map the host ids, but it seems
to be the most intuitive given that xhost is required to allow
the container application to connect to X11; in the case of userns
host it will be container uid 0 -> uid 0 host but feel free to use
this at your own discretion to say the least it will help some
people who don't have a node/npm that's worth messing with installed
on their local machine to use this as well.

update to Dockerfile, use non-root user for operation in userns host mode

fix chown to use /home/iohzrd

fix data persistence volume and homedir/path/permission issue to agree with electron